Fix cyclic transitive reduction bug in graphs with 1-vertex SCCs
An SCC has either a single vertex and no edges or at least 3 vertices and 3
edges for a directed graph. Before the bugfix the algorithm was faulty insofar
as it always assumed that an SCC has 3 edges and even asserted for it. When
testing with another application, I noticed, then fixed the bug and added a test
case for it.